Split KlumAnnotationsValidator into separate instances for fields and methods
